+++ app/src/renderer/components/staking/PageValidator.vue
@@ -1,8 +1,6 @@
 <template>
   <tm-page data-title="Validator">
-    <tm-data-loading v-if="delegates.loading" />
     <tm-data-error v-if="delegates.loaded && !validator" />
-
     <template v-else>
       <div class="page-profile__header page-profile__section">
         <div class="row">
@@ -189,7 +187,6 @@ import { mapGetters } from "vuex"
 import { percent, pretty, atoms, full } from "scripts/num"
 import TmBtn from "common/TmBtn"
 import TmModal from "common/TmModal"
-import TmDataLoading from "common/TmDataLoading"
 import TmDataError from "common/TmDataError"
 import { shortAddress, ratToBigNumber } from "scripts/common"
 import DelegationModal from "staking/DelegationModal"
@@ -205,7 +202,6 @@ export default {
     UndelegationModal,
     TmBtn,
     TmModal,
-    TmDataLoading,
     TmDataError,
     TmPage
   },
@@ -265,7 +261,7 @@ export default {
     },
     myDelegation() {
       const { bondDenom, myBond } = this
-      const myDelegation = full(atoms(myBond))
+      const myDelegation = full(myBond)
       const myDelegationString = `${myDelegation} ${bondDenom}`
       return Number(myBond) === 0 ? `--` : myDelegationString
     },
@@ -307,16 +303,14 @@ export default {
     },
     rewards() {
       const { session, bondDenom, distribution, validator } = this
-      if (!session.signedIn) { 
-        return null 
+      if (!session.signedIn) {
+        return null
       }
 
-      const validatorRewards = distribution.rewards[
-        validator.operator_address
-      ]
-      const amount = validatorRewards ? full(
-        atoms(validatorRewards[bondDenom]) || 0
-      ) : null
+      const validatorRewards = distribution.rewards[validator.operator_address]
+      const amount = validatorRewards
+        ? full(atoms(validatorRewards[bondDenom]) || 0)
+        : null
 
       if (amount) {
         return `${amount} ${bondDenom}`
@@ -334,7 +328,7 @@ export default {
     },
     lastHeader: {
       immediate: true,
-      handler(){
+      handler() {
         if (this.session.signedIn) {
           this.$store.dispatch(
             `getRewardsFromValidator`,